split bases
