+++ lib/asciidoctor/rfc/v3/base.rb
@@ -1,0 +1,249 @@
+module Asciidoctor
+  module RFC::V3
+    module Base
+      # Syntax:
+      #   =Title
+      #   Author
+      #   :ipr
+      #   :obsoletes
+      #   :updates
+      #   :submissionType
+      #   :indexInclude
+      #   :iprExtract
+      #   :sortRefs
+      #   :symRefs
+      #   :tocInclude
+      #
+      #   ABSTRACT
+      #
+      #   NOTEs
+      #
+      #   ==first title
+      #   CONTENT
+      #
+      #   [bibliography] # start of back matter
+      #   == Bibliography
+      #
+      #   [appendix] # start of back matter if not already started
+      #   == Appendix
+      def document(node)
+        result = []
+        result << '<?xml version="1.0" encoding="UTF-8"?>'
+        ipr = get_header_attribute node, "ipr"
+        obsoletes = get_header_attribute node, "obsoletes"
+        updates = get_header_attribute node, "updates"
+        indexInclude = get_header_attribute node, "indexInclude"
+        iprExtract = get_header_attribute node, "iprExtract"
+        sortRefs = get_header_attribute node, "sortRefs"
+        symRefs = get_header_attribute node, "symRefs"
+        tocInclude = get_header_attribute node, "tocInclude"
+        submissionType = get_header_attribute node, "submissionType", "IETF"
+        t = Time.now.getutc
+        preptime = set_header_attribute "preptime",
+          sprintf("%04d-%02d-%02dT%02d:%02d:%02dZ", t.year, t.month, t.day, t.hour, t.min, t.sec)
+        version = set_header_attribute "version", "3"
+        result << %(<rfc#{document_ns_attributes node}#{ipr}#{obsoletes}#{updates}#{preptime}
+        #{version}#{submissionType}#{indexInclude}#{iprExtract}#{sortRefs}#{symRefs}#{tocInclude}>)
+        result << (link node)
+        result << (front node)
+        result << "</front><middle1>"
+        result << node.content if node.blocks?
+        result << ($seen_back_matter ? "</back>" : "</middle>")
+        result << "</rfc>"
+
+        # <middle> needs to move after preamble
+        result = result.flatten
+        if result.any? { |e| e =~ /<\/front><middle>/ } and result.any? { |e| e =~ /<\/front><middle1>/ }
+          result = result.reject { |e| e =~ /<\/front><middle1>/ }
+        else
+          result = result.map { |e| e =~ /<\/front><middle1>/ ? "</front><middle>" : e }
+        end
+
+        result * "\n"
+      end
+
+      # Syntax:
+      #   = Title
+      #   Author
+      #   :link href,href rel
+      def link(node)
+        result = []
+        links = node.attr "link"
+        return result if links.nil?
+        links.split(/,/).each do |l|
+          m = /^(\S+)\s+(\S+)$/.match(l)
+          if m.nil?
+            href = set_header_attribute "href", l
+            result << "<link#{href}/>"
+          else
+            href = set_header_attribute "href", m[1]
+            rel = set_header_attribute "rel", m[2]
+            result << "<link#{href}#{rel}/>"
+          end
+        end
+        result
+      end
+
+      def inline_break(node)
+        %(#{node.text}<br/>)
+      end
+
+      def inline_quoted(node)
+        case node.type
+        when :emphasis
+          "<em>#{node.text}</em>"
+        when :strong
+          "<strong>#{node.text}</strong>"
+        when :monospaced
+          "<tt>#{node.text}</tt>"
+        when :double
+          "\"#{node.text}\""
+        when :single
+          "'#{node.text}'"
+        when :superscript
+          "<sup>#{node.text}</sup>"
+        when :subscript
+          "<sub>#{node.text}</sub>"
+        else
+          # [bcp14]#MUST NOT#
+          if node.role == "bcp14"
+            "<bcp14>#{node.text}</bcp14>"
+          else
+            node.text
+          end
+        end
+      end
+
+      # Syntax:
+      #   [[id]]
+      #   [keepWithNext=true,keepWithPrevious=true] (optional)
+      #   Text
+      def paragraph(node)
+        result = []
+        if node.parent.context == :preamble and not $seen_abstract
+          $seen_abstract = true
+          result << "<abstract>"
+        end
+        id = set_header_attribute "anchor", node.id
+        keepWithNext = get_header_attribute node, "keepWithNext"
+        keepWithPrevious = get_header_attribute node, "keepWithPrevious"
+        result << "<t#{id}#{keepWithNext}#{keepWithPrevious}>#{node.content}</t>"
+        result
+      end
+
+      # Syntax:
+      #   [[id]]
+      #   [removeInRFC=true,toc=include|exclude|default] (optional)
+      #   == title
+      #   Content
+      #
+      #   [[id]]
+      #   [bibliography]
+      #   == Normative|Informative References
+      #   * [[[ref1]]] Ref [must provide references as list]
+      #   * [[[ref2]]] Ref
+      def section(node)
+        result = []
+        if node.attr("style") == "bibliography"
+          $xreftext = {}
+          $processing_reflist = true
+          id = set_header_attribute "anchor", node.id
+          result << "<references#{id}>"
+          result << "<name>#{node.title}</name>" unless node.title.nil?
+          # require that references be given in a ulist
+          node.blocks.each { |b| result << reflist(b) }
+          result << "</references>"
+          unless $xreftext.empty?
+            result.unshift($xreftext.keys.map { |k| %(<displayreference target="#{k}" to="#{$xreftext[k]}"/>) })
+          end
+          result = result.unshift("</middle><back>") unless $seen_back_matter
+          $processing_reflist = false
+          $seen_back_matter = true
+        else
+          id = set_header_attribute "anchor", node.id
+          removeInRFC = get_header_attribute node, "removeInRFC"
+          toc = get_header_attribute node, "toc"
+          numbered = set_header_attribute "numbered", node.attr?("sectnums")
+          if node.attr("style") == "appendix"
+            result << "</middle><back>" unless $seen_back_matter
+            $seen_back_matter = true
+          end
+          result << "<section#{id}#{removeInRFC}#{toc}#{numbered}>"
+          result << "<name>#{node.title}</name>" unless node.title.nil?
+          result << node.content
+          result << "</section>"
+        end
+        result
+      end
+
+      # Syntax:
+      #   [[id]]
+      #   .Title
+      #   |===
+      #   |col | col
+      #   |===
+      def table(node)
+        has_body = false
+        result = []
+        id = set_header_attribute "anchor", node.id
+        result << %(<table#{id}>)
+        result << %(<name>#{node.title}</name>) if node.title?
+        # TODO iref belongs here
+        [:head, :body, :foot].reject { |tblsec| node.rows[tblsec].empty? }.each do |tblsec|
+          has_body = true if tblsec == :body
+          # id = set_header_attribute "anchor", tblsec.id
+          # not supported
+          result << %(<t#{tblsec}>)
+          node.rows[tblsec].each_with_index do |row, i|
+            # id not supported on row
+            result << "<tr>"
+            rowlength = 0
+            result1 = []
+            row.each do |cell|
+              id = set_header_attribute "anchor", cell.id
+              colspan_attribute = set_header_attribute "colspan", cell.colspan
+              rowspan_attribute = set_header_attribute "rowspan", cell.rowspan
+              align = set_header_attribute("align", cell.attr("halign"))
+              cell_tag_name = (tblsec == :head || cell.style == :header ? "th" : "td")
+              entry_start = %(<#{cell_tag_name}#{colspan_attribute}#{rowspan_attribute}#{id}#{align}>)
+              cell_content = cell.text
+              rowlength += cell_content.size
+              result1 << %(#{entry_start}#{cell_content}</#{cell_tag_name}>)
+            end
+            result << result1
+            if rowlength > 72
+              warn "asciidoctor: WARNING: row #{i} of table (count including header rows) is longer than 72 ascii characters:\n#{result1}"
+            end
+            result << "</tr>"
+          end
+          result << %(</t#{tblsec}>)
+        end
+        result << "</table>"
+
+        warn "asciidoctor: WARNING: tables must have at least one body row" unless has_body
+        result
+      end
+
+      # Syntax:
+      #   [[id]]
+      #   .Name
+      #   [link=xxx,align=left|center|right,alt=alt_text,type]
+      #   image::filename[]
+      # @note ignoring width, height attributes
+      def image(node)
+        result = []
+        result << "<figure>" if node.parent.context != :example
+        id = set_header_attribute "anchor", node.id
+        align = get_header_attribute node, "align"
+        alt = set_header_attribute "alt", node.alt
+        link = (node.image_uri node.target)
+        src = set_header_attribute node, "src", link
+        type = set_header_attribute node, "type", link =~ /\.svg$/ ? "svg" : "binary-art"
+        name = nil
+        result << "<artwork#{id}#{name}#{align}#{alt}#{type}#{src}/>"
+        result << "</figure>" if node.parent.context != :example
+        result
+      end
+    end
+  end
+end
